fix: all
- Correções de nomes de arquivos diversos. repositórios, e comentários.

- Adicionado STATUS na struct de usuários.
