+++ README.md
@@ -1,2 +1,31 @@
-# intranet-fieldservice-atento-api
-Intranet desenvolvida de forma não oficial para auxilio à equipe de Field Service da Atento Brasil.
+# INTRANET FIELD SERVICE - API
+A Intranet do Field Service é um sistema desenvolvido de forma não oficial para auxilio e centralização de informações utilizadas na área de TI da Atento Brasil.
+
+### Funcionalidades
+- Página Inicial com centralização de ferramentas/links úteis. 
+- Mapa Operacional (Informações úteis das operações) 
+- Base de Conhecimento
+- Integrações com Sistemas de Restrição de Acesso a Internet e Gestão de Ativos.
+
+## 💻 Pré-requisitos
+Antes de começar, verifique se você atendeu aos seguintes requisitos:
+
+* Versão mais recente do `Golang`
+
+## 🚀 Rodando a API da INTRANET
+
+Para rodar a API, siga as etapas listadas abaixo: 
+
+- Importar os scripts presentes na pasta "sql" no banco de dados MYSQL ou MARIADB.
+- Defina as credenciais de acesso ao banco de dados no arquivo .env.
+- Compile e Execute a API com os seguintes comandos:
+  
+```
+go build api
+api.exe
+```
+
+## ☕ Documentação da API
+```
+(soon)
+```
