+++ go.mod
@@ -16,4 +16,6 @@ require (
 	github.com/abdfnx/gosh v0.3.9 // indirect
 	github.com/jasonlvhit/gocron v0.0.1 // indirect
 	github.com/robfig/cron/v3 v3.0.0 // indirect
+	github.com/szyhf/go-convert v1.0.0 // indirect
+	github.com/szyhf/go-excel v1.5.1 // indirect
 )
