+++ src/controllers/usuarios.go
@@ -209,33 +209,3 @@ func BuscarUsuario(w http.ResponseWriter, r *http.Request) {
 
 	respostas.JSON(w, http.StatusOK, usuario)
 }
-
-// DeletarSessionDB deleta a sessão armazenada no banco
-func DeletarSessionDB(w http.ResponseWriter, r *http.Request) {
-
-	parametros := mux.Vars(r)
-	userID, erro := strconv.ParseUint(parametros["userId"], 10, 64)
-	if erro != nil {
-		respostas.Erro(w, http.StatusBadRequest, erro)
-		return
-	}
-
-	db, erro := banco.Conectar()
-	if erro != nil {
-		respostas.Erro(w, http.StatusInternalServerError, erro)
-		return
-	}
-	defer db.Close()
-
-	repositorioD := repositorios.NovoRepositorioDeSessions(db)
-	sessionArmazenadaDB, erro := repositorioD.BuscarPorID(userID)
-	if erro != nil {
-		return
-	}
-
-	repositorio := repositorios.NovoRepositorioDeSessions(db)
-	if erro = repositorio.DeletarSession(sessionArmazenadaDB.ID); erro != nil {
-		respostas.Erro(w, http.StatusInternalServerError, erro)
-		return
-	}
-}
