+++ src/modelos/Dac.go
@@ -1,0 +1,7 @@
+package modelos
+
+// Dac representa um dac cadastrado em sistema
+type Dac struct {
+	IDDAC uint64 `json:"iddac,omitempty"`
+	NOME  string `json:"nome,omitempty"`
+}
