+++ src/modelos/Dominio.go
@@ -1,0 +1,7 @@
+package modelos
+
+// Dominio representa um domínio cadastrado no sistema
+type Dominio struct {
+	IDDOMINIO uint64 `json:"iddominio,omitempty"`
+	NOME      string `json:"nome,omitempty"`
+}
