+++ src/modelos/Log.go
@@ -2,6 +2,7 @@ package modelos
 
 import "time"
 
+//Logs representa um log cadastrado no sistema
 type Logs struct {
 	Usuario Usuario
 	ACTION  string    `json:"action,omitempty"`
