+++ src/modelos/Session.go
@@ -2,6 +2,7 @@ package modelos
 
 import "time"
 
+//Session representa o registro de um sessão logada.
 type Session struct {
 	ID                uint64 `json:"id,omitempty"`
 	Usuario           Usuario
