+++ src/modelos/Usuario.go
@@ -12,7 +12,7 @@ import (
 // Usuario representa um usuário cadastrado no sistema
 type Usuario struct {
 	IDUSUARIO              uint64    `json:"idusuario,omitempty"`
-	NOME                   string    `json:"nome,omitempty"`
+	NOME                   string    `json:"nome,omitempty""`
 	LOGIN_NT               string    `json:"login_nt,omitempty"`
 	RE                     string    `json:"re,omitempty"`
 	CARGO                  string    `json:"cargo,omitempty"`
@@ -28,6 +28,7 @@ type Usuario struct {
 	ID_SITE                string    `json:"id_site,omitempty"`
 	DATA_CRIACAO           time.Time `json:"data_criacao,omitempty"`
 	Site                   Site
+	STATUS                 string `json:"status,omitempty"`
 }
 
 // Struct Senha representa o formato da requisição de alteração de senha
