+++ src/repositorios/usuario.go
@@ -19,14 +19,14 @@ func NovoRepositorioDeUsuarios(db *sql.DB) *Usuarios {
 // Criar insere um novo usuário no banco de dados
 func (repositorio Usuarios) CriarUsuario(usuario modelos.Usuario) (uint64, error) {
 	statement, erro := repositorio.db.Prepare(
-		"INSERT INTO USUARIOS(NOME, LOGIN_NT, RE, CARGO, EMAIL, SENHA, V_USUARIOS, V_BDC_POSTS, V_BDC_ADM, V_IMDB, V_GSA, V_MAPA_OPERACIONAL, V_MAPA_OPERACIONAL_ADM, ID_SITE) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+		"INSERT INTO USUARIOS(NOME, LOGIN_NT, RE, CARGO, EMAIL, SENHA, V_USUARIOS, V_BDC_POSTS, V_BDC_ADM, V_IMDB, V_GSA, V_MAPA_OPERACIONAL, V_MAPA_OPERACIONAL_ADM, ID_SITE, STATUS) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
 	)
 	if erro != nil {
 		return 0, erro
 	}
 	defer statement.Close()
 
-	resultado, erro := statement.Exec(usuario.NOME, usuario.LOGIN_NT, usuario.RE, usuario.CARGO, usuario.EMAIL, usuario.SENHA, usuario.V_USUARIOS, usuario.V_BDC_POSTS, usuario.V_BDC_ADM, usuario.V_IMDB, usuario.V_GSA, usuario.V_MAPA_OPERACIONAL, usuario.V_MAPA_OPERACIONAL_ADM, usuario.ID_SITE)
+	resultado, erro := statement.Exec(usuario.NOME, usuario.LOGIN_NT, usuario.RE, usuario.CARGO, usuario.EMAIL, usuario.SENHA, usuario.V_USUARIOS, usuario.V_BDC_POSTS, usuario.V_BDC_ADM, usuario.V_IMDB, usuario.V_GSA, usuario.V_MAPA_OPERACIONAL, usuario.V_MAPA_OPERACIONAL_ADM, usuario.ID_SITE, usuario.STATUS)
 	if erro != nil {
 		return 0, erro
 	}
@@ -47,7 +47,7 @@ func (repositorio Usuarios) Buscar(nomeOuLogin string) ([]modelos.Usuario, error
 	nomeOuLogin = fmt.Sprintf("%%%s%%", nomeOuLogin)
 
 	linhas, erro := repositorio.db.Query(
-		"SELECT U.IDUSUARIO, U.NOME, U.RE, U.LOGIN_NT, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, U.ID_SITE, S.IDSITE, S.NOME FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE WHERE U.NOME LIKE ? or U.LOGIN_NT LIKE ? ORDER BY U.IDUSUARIO DESC", nomeOuLogin, nomeOuLogin)
+		"SELECT U.IDUSUARIO, U.NOME, U.RE, U.LOGIN_NT, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, U.ID_SITE, U.STATUS, S.IDSITE, S.NOME FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE WHERE U.NOME LIKE ? or U.LOGIN_NT LIKE ? ORDER BY U.IDUSUARIO DESC", nomeOuLogin, nomeOuLogin)
 
 	if erro != nil {
 		return nil, erro
@@ -74,6 +74,7 @@ func (repositorio Usuarios) Buscar(nomeOuLogin string) ([]modelos.Usuario, error
 			&usuario.V_MAPA_OPERACIONAL,
 			&usuario.V_MAPA_OPERACIONAL_ADM,
 			&usuario.ID_SITE,
+			&usuario.STATUS,
 			&usuario.Site.IDSITE,
 			&usuario.Site.NOME,
 		); erro != nil {
@@ -89,7 +90,7 @@ func (repositorio Usuarios) Buscar(nomeOuLogin string) ([]modelos.Usuario, error
 // BuscarPorID traz um usuário do banco de dados filtrado pelo id
 func (repositorio Usuarios) BuscarPorID(ID uint64) (modelos.Usuario, error) {
 	linhas, erro := repositorio.db.Query(
-		"SELECT U.IDUSUARIO, U.NOME, U.RE, U.LOGIN_NT, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, U.ID_SITE, S.IDSITE, S.NOME FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE WHERE U.IDUSUARIO = ?", ID)
+		"SELECT U.IDUSUARIO, U.NOME, U.RE, U.LOGIN_NT, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, U.ID_SITE, U.STATUS, S.IDSITE, S.NOME FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE WHERE U.IDUSUARIO = ?", ID)
 	if erro != nil {
 		return modelos.Usuario{}, erro
 	}
@@ -113,6 +114,7 @@ func (repositorio Usuarios) BuscarPorID(ID uint64) (modelos.Usuario, error) {
 			&usuario.V_MAPA_OPERACIONAL,
 			&usuario.V_MAPA_OPERACIONAL_ADM,
 			&usuario.ID_SITE,
+			&usuario.STATUS,
 			&usuario.Site.IDSITE,
 			&usuario.Site.NOME,
 		); erro != nil {
@@ -126,14 +128,14 @@ func (repositorio Usuarios) BuscarPorID(ID uint64) (modelos.Usuario, error) {
 // Atualizar altera as informações de um usuário no banco de dados
 func (repositorio Usuarios) Atualizar(ID uint64, usuario modelos.Usuario) error {
 	statement, erro := repositorio.db.Prepare(
-		"UPDATE USUARIOS SET NOME = ?, RE = ?, LOGIN_NT = ?, CARGO = ?, EMAIL = ?, V_USUARIOS = ?, V_BDC_POSTS = ?, V_BDC_ADM = ?, V_IMDB = ?, V_GSA = ?, V_MAPA_OPERACIONAL = ?, V_MAPA_OPERACIONAL_ADM = ?, ID_SITE = ? WHERE IDUSUARIO = ?",
+		"UPDATE USUARIOS SET NOME = ?, RE = ?, LOGIN_NT = ?, CARGO = ?, EMAIL = ?, V_USUARIOS = ?, V_BDC_POSTS = ?, V_BDC_ADM = ?, V_IMDB = ?, V_GSA = ?, V_MAPA_OPERACIONAL = ?, V_MAPA_OPERACIONAL_ADM = ?, ID_SITE = ?, STATUS = ? WHERE IDUSUARIO = ?",
 	)
 	if erro != nil {
 		return erro
 	}
 	defer statement.Close()
 
-	if _, erro = statement.Exec(usuario.NOME, usuario.RE, usuario.LOGIN_NT, usuario.CARGO, usuario.EMAIL, usuario.V_USUARIOS, usuario.V_BDC_POSTS, usuario.V_BDC_ADM, usuario.V_IMDB, usuario.V_GSA, usuario.V_MAPA_OPERACIONAL, usuario.V_MAPA_OPERACIONAL_ADM, usuario.ID_SITE, ID); erro != nil {
+	if _, erro = statement.Exec(usuario.NOME, usuario.RE, usuario.LOGIN_NT, usuario.CARGO, usuario.EMAIL, usuario.V_USUARIOS, usuario.V_BDC_POSTS, usuario.V_BDC_ADM, usuario.V_IMDB, usuario.V_GSA, usuario.V_MAPA_OPERACIONAL, usuario.V_MAPA_OPERACIONAL_ADM, usuario.ID_SITE, usuario.STATUS, ID); erro != nil {
 		return erro
 	}
 
@@ -157,7 +159,7 @@ func (repositorio Usuarios) Deletar(ID uint64) error {
 
 // BuscarPorEmail busca um usuário por email e retorna o seu id e senha com hash
 func (repositorio Usuarios) BuscarPorEmail(email string) (modelos.Usuario, error) {
-	linha, erro := repositorio.db.Query("SELECT U.IDUSUARIO, U.LOGIN_NT, U.NOME, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, S.IDSITE, S.NOME, U.SENHA FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE = U.IDUSUARIO WHERE EMAIL = ?", email)
+	linha, erro := repositorio.db.Query("SELECT U.IDUSUARIO, U.LOGIN_NT, U.NOME, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, U.STATUS, S.IDSITE, S.NOME, U.SENHA FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE = U.IDUSUARIO WHERE EMAIL = ?", email)
 	if erro != nil {
 		return modelos.Usuario{}, erro
 	}
@@ -166,7 +168,7 @@ func (repositorio Usuarios) BuscarPorEmail(email string) (modelos.Usuario, error
 	var usuario modelos.Usuario
 
 	if linha.Next() {
-		if erro = linha.Scan(&usuario.IDUSUARIO, &usuario.LOGIN_NT, &usuario.NOME, &usuario.CARGO, &usuario.EMAIL, &usuario.V_USUARIOS, &usuario.V_BDC_POSTS, &usuario.V_BDC_ADM, &usuario.V_IMDB, &usuario.V_GSA, &usuario.V_MAPA_OPERACIONAL, &usuario.V_MAPA_OPERACIONAL_ADM, &usuario.Site.IDSITE, &usuario.Site.NOME, &usuario.SENHA); erro != nil {
+		if erro = linha.Scan(&usuario.IDUSUARIO, &usuario.LOGIN_NT, &usuario.NOME, &usuario.CARGO, &usuario.EMAIL, &usuario.V_USUARIOS, &usuario.V_BDC_POSTS, &usuario.V_BDC_ADM, &usuario.V_IMDB, &usuario.V_GSA, &usuario.V_MAPA_OPERACIONAL, &usuario.V_MAPA_OPERACIONAL_ADM, &usuario.STATUS, &usuario.Site.IDSITE, &usuario.Site.NOME, &usuario.SENHA); erro != nil {
 			return modelos.Usuario{}, erro
 		}
 	}
@@ -177,7 +179,7 @@ func (repositorio Usuarios) BuscarPorEmail(email string) (modelos.Usuario, error
 
 // BuscarPorLogin busca um usuário pelo Login
 func (repositorio Usuarios) BuscarPorLogin(LOGIN_NT string) (modelos.Usuario, error) {
-	linha, erro := repositorio.db.Query("SELECT U.IDUSUARIO, U.LOGIN_NT, U.RE, U.NOME, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, S.IDSITE, S.NOME, U.SENHA FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE WHERE U.LOGIN_NT = ?", LOGIN_NT)
+	linha, erro := repositorio.db.Query("SELECT U.IDUSUARIO, U.LOGIN_NT, U.RE, U.NOME, U.CARGO, U.EMAIL, U.V_USUARIOS, U.V_BDC_POSTS, U.V_BDC_ADM, U.V_IMDB, U.V_GSA, U.V_MAPA_OPERACIONAL, U.V_MAPA_OPERACIONAL_ADM, U.STATUS, S.IDSITE, S.NOME, U.SENHA FROM USUARIOS U INNER JOIN SITES S ON S.IDSITE = U.ID_SITE WHERE U.LOGIN_NT = ?", LOGIN_NT)
 	if erro != nil {
 		return modelos.Usuario{}, erro
 	}
@@ -186,7 +188,7 @@ func (repositorio Usuarios) BuscarPorLogin(LOGIN_NT string) (modelos.Usuario, er
 	var usuario modelos.Usuario
 
 	if linha.Next() {
-		if erro = linha.Scan(&usuario.IDUSUARIO, &usuario.LOGIN_NT, &usuario.RE, &usuario.NOME, &usuario.CARGO, &usuario.EMAIL, &usuario.V_USUARIOS, &usuario.V_BDC_POSTS, &usuario.V_BDC_ADM, &usuario.V_IMDB, &usuario.V_GSA, &usuario.V_MAPA_OPERACIONAL, &usuario.V_MAPA_OPERACIONAL_ADM, &usuario.Site.IDSITE, &usuario.Site.NOME, &usuario.SENHA); erro != nil {
+		if erro = linha.Scan(&usuario.IDUSUARIO, &usuario.LOGIN_NT, &usuario.RE, &usuario.NOME, &usuario.CARGO, &usuario.EMAIL, &usuario.V_USUARIOS, &usuario.V_BDC_POSTS, &usuario.V_BDC_ADM, &usuario.V_IMDB, &usuario.V_GSA, &usuario.V_MAPA_OPERACIONAL, &usuario.V_MAPA_OPERACIONAL_ADM, &usuario.STATUS, &usuario.Site.IDSITE, &usuario.Site.NOME, &usuario.SENHA); erro != nil {
 			return modelos.Usuario{}, erro
 		}
 	}
